Split solution into modules
